Create search_file_and_move
This script uses constants to improve reusability and readability, the constants are "SEARCH_DIRECTORY", "FILENAME", and "DESTINATION_DIRECTORY", these constants can be changed easily in the script to match different requirements.

The script uses the find command to search all directories in the $SEARCH_DIRECTORY for a file named $FILENAME, the -print option is used to print the full path of the file, and -quit option is used to stop the search after the first file is found. The result is stored in a variable "path".

The if statement is used to check if the variable "path" is not empty and move the file if it is not empty. If the variable "path" is empty it means that the file is not found, the script will display the message "File not found".

Please note that this script will search through all the directories in the $SEARCH_DIRECTORY drive, it may take a while to run depending on the amount of files and directories on the drive. Also, it is always recommended to test the script in a safe environment before running it in production, and make sure that you have permission to access the directories that you are searching through.